+++ nail/CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- added CLI param `--mmseqs-path` 
+
 
 ## [0.5.0] - 2026-3-19